Project organization.
EAGLe is being split into two (well.. three?) subprojects.

These subprojects are:
SHPSCore which handles SHPS images
and eagle-cli which is the CLI interface for EAGLe.

SHPSCore is intended to be used by third party projects, so a section in the README has also been added to detail how to properly import SHPSCore.