[Core] Library restructuring
In preparation for mcommon to be obliterated and bigfile/makecbxs being rolled in,
there has been a lot of work torwards rethinking how to do structures and reading/writing of them.
Much of the core and shps libraries will be replaced with better code, which will also when the time comes be adaptable to writing.

The GUI and all of the old code still compiles with this commit, however,
that code won't be there forever...